Move `LogHandler` and `StageResultHandler` to plugin-util.
