+++ plugin/src/main/java/io/jenkins/plugins/analysis/core/steps/AnnotatedReport.java
@@ -23,7 +23,7 @@ import io.jenkins.plugins.forensics.miner.RepositoryStatistics;
  *
  * @author Ullrich Hafner
  */
-public class AnnotatedReport implements Serializable {
+public class    AnnotatedReport implements Serializable {
     private static final long serialVersionUID = -4797152016409014028L;
 
     private final String id;
