+++ plugin/src/main/java/io/jenkins/plugins/analysis/core/steps/IssuesPublisher.java
@@ -24,13 +24,13 @@ import io.jenkins.plugins.analysis.core.model.ResetReferenceAction;
 import io.jenkins.plugins.analysis.core.model.ResultAction;
 import io.jenkins.plugins.analysis.core.model.ResultSelector;
 import io.jenkins.plugins.analysis.core.util.HealthDescriptor;
-import io.jenkins.plugins.analysis.core.util.LogHandler;
 import io.jenkins.plugins.analysis.core.util.QualityGateEvaluator;
 import io.jenkins.plugins.analysis.core.util.QualityGateStatus;
-import io.jenkins.plugins.analysis.core.util.StageResultHandler;
 import io.jenkins.plugins.analysis.core.util.TrendChartType;
 import io.jenkins.plugins.forensics.reference.ReferenceFinder;
 import io.jenkins.plugins.util.JenkinsFacade;
+import io.jenkins.plugins.util.LogHandler;
+import io.jenkins.plugins.util.StageResultHandler;
 
 import static io.jenkins.plugins.analysis.core.model.AnalysisHistory.JobResultEvaluationMode.*;
 import static io.jenkins.plugins.analysis.core.model.AnalysisHistory.QualityGateEvaluationMode.*;
@@ -119,7 +119,8 @@ class IssuesPublisher {
         }
 
         issues.logInfo("Attaching ResultAction with ID '%s' to build '%s'.", getId(), run);
-        logger.log(issues);
+        logger.logInfoMessages(issues.getInfoMessages());
+        logger.logErrorMessages(issues.getErrorMessages());
 
         AnalysisResult result = new AnalysisHistory(run, selector).getResult()
                 .map(previous -> new AnalysisResult(run, getId(), deltaReport, report.getBlames(),
